server/requirements.txt
flask
flask-cors
faster-whisper
google-generativeai
python-dotenv
Werkzeug
